% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{get.para}
\alias{getResults}
\alias{getResults.Boot}
\alias{getResults.BootV1}
\alias{getResults.CV}
\title{getResults}
\usage{
getResults.CV(res)

getResults.Boot(res, alpha = 0.05)

getResults.BootV1(res, alpha = 0.05)

get.para(res)
}
\arguments{
\item{res}{predictions}

\item{alpha}{significance level}
}
\value{
predictions
}
\description{
Extract prediction results
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}

