% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preProcessData.R
\name{preProcessData}
\alias{preProcessData}
\title{preProcessData}
\usage{

  preProcessData(HighRiskMedicaid.data.file.name = "AnalyticFileForNilay_090814.dta")
}
\arguments{
\item{HighRiskMedicaid.data.file.name}{data file name}
}
\value{
training data, lsit of model domains, respons variables and list of fomulas.
}
\description{
Set up data
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

