% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RunModels.R
\name{RunModels}
\alias{RunHighRiskMedicaid.Boot}
\alias{RunHighRiskMedicaid.CV}
\alias{RunModels}
\title{RunModels}
\usage{
RunHighRiskMedicaid.CV(classifier, cv = 2, XY.dat, resp.vars, rhs.vars.list,
  para, opt.para = FALSE, do.STL = TRUE, do.MTL = FALSE,
  MTL.opt.para = FALSE, task.type = NULL, seed = 12345678)

RunHighRiskMedicaid.Boot(classifier, nBoots = 5, XY.dat, resp.vars,
  rhs.vars.list, MTL.para = NULL, STL.para = NULL, do.STL = TRUE,
  do.MTL = FALSE, task.type, opt.para = FALSE, MTL.opt.para = FALSE)
}
\arguments{
\item{classifier}{character list of classification models. See names(TrainAllModels()).}

\item{cv,nBoots}{number of folds, bootsraps}

\item{XY.dat}{data matrix}

\item{resp.vars}{reponse variables : total expenditure, frequent ER visits,
frequent hopitalization}

\item{rhs.vars.list}{named list of predictor variables for each incremental
model}

\item{para,STL.para,MTL.para}{named list of model parameters}

\item{opt.para,MTL.opt.para}{logical for parameter tuning. defaul to FALSE}

\item{do.STL,do.MTL}{logicals}

\item{task.type}{parameter for MTL}

\item{seed}{random seed}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
list of functions for training various algorithms
}
\description{
Run, cross-validation or bootstrap in parallel
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}

