% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RUSBoosModels.R
\name{RUSBoostModels}
\alias{RUSBoostELR}
\alias{RUSBoostELR.default}
\alias{RUSBoostELR.formula}
\alias{RUSBoostGBM}
\alias{RUSBoostGLM}
\alias{RUSBoostGLM.default}
\alias{RUSBoostGLM.formula}
\alias{RUSBoostModels}
\alias{predict.RUSBoostELR}
\alias{predict.RUSBoostGBM}
\alias{predict.RUSBoostGLM}
\title{RUSBoostModels}
\usage{
RUSBoostELR(x, ...)

\method{RUSBoostELR}{default}(x, ...)

\method{RUSBoostELR}{formula}(form, data, boot = FALSE, iters = 100,
  coeflearn = "Breiman", sampleFraction, para, idx, ...)

\method{predict}{RUSBoostELR}(object, newdata,
  newmfinal = length(object$trees), ...)

RUSBoostGLM(x, ...)

\method{RUSBoostGLM}{default}(x, ...)

\method{RUSBoostGLM}{formula}(formula, data, boot = FALSE, iters = 100,
  coeflearn = "Breiman", sampleFraction, idx, ...)

\method{predict}{RUSBoostGLM}(object, newdata,
  newmfinal = length(object$trees), ...)

RUSBoostGBM(formula, data, boot = FALSE, iters = 100,
  coeflearn = "Breiman", sampleFraction, para, idx)

RUSBoostGBM(formula, data, boot = FALSE, iters = 100,
  coeflearn = "Breiman", sampleFraction, para, idx)

\method{predict}{RUSBoostGBM}(object, newdata,
  newmfinal = length(object$trees), ...)
}
\arguments{
\item{form,formula}{A formula}

\item{data,x}{data matrix}

\item{boot}{logical}

\item{iters}{number of bootstrap replicates}

\item{coeflearn}{from \code{adabag}}

\item{sampleFraction}{undersampling ratio}

\item{para}{A named parameter list of model parameters}

\item{idx}{indices of majority class}

\item{object}{trained model of appropriate class}

\item{newdata,newmfinal}{newdata}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Trained undersample bootstrap models
}
\description{
Randon undersampling with boosting for various learners.
Learner does not have to be a weak learner as in standard boosting
algorithm
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

