% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{CreateDummy}
\alias{Split}
\alias{SplitCrossVal}
\alias{SplitCrossVal2}
\alias{TrainValTestSplit}
\alias{Utils}
\alias{collect.garbage}
\alias{my.tryCatch}
\title{Utility functions.}
\usage{
TrainValTestSplit(D.dat, test = TRUE)

Split(Nt, K)

SplitCrossVal(D.dat, k, ix.cv, train.propo = 0.85, test = TRUE)

SplitCrossVal2(D.dat, k, ix.cv, val.propo = 0.15)

collect.garbage()

CreateDummy(D.dat)

my.tryCatch(code)
}
\arguments{
\item{D.dat}{A matix/data frame}

\item{test}{logical: should test set be created?}

\item{Nt}{Sample size}

\item{K,k}{number of cross-validation and cross-validation number}

\item{ix.cv}{cross-validation indices}

\item{train.propo,val.propo}{training set proportion}

\item{code}{function or expression}
}
\value{
training, validation, test data sets or K-fold cross-validation data sets.
}
\description{
Some useful miscellaneous functions
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

